[IMP] render: connectivity of render server
After getting no requests for 15 minutes on the Bot Instance
hosted on Render the instance gets spun down and it won't be
able to respond to any requests from users.

The other thing is that render uses `web dyno` which requires
a web server to be actively running.

So, to resolve this we start an Interval which will run the
`keepAwake` function which will ping the Instance at after
10 minutes.

We create a HTTP server for the server issues.